+++ docs/guides/message-debugging.mdx
@@ -52,6 +52,17 @@ import CosmosMessageDelivered from "@site/src/components/CosmosMessageDelivered"
 </details>
 
 <details>
+  <summary>**Debugging a message to a Sealevel chain (e.g. Solana, Eclipse)**</summary>
+  <div>
+    At the moment, the [Hyperlane Explorer](https://explorer.hyperlane.xyz) doesn't support non-EVM chains.
+
+    To check if your message has been delivered, select the destination chain and input your message ID below:
+
+    <CosmosMessageDelivered />
+  </div>
+</details>
+
+<details>
   <summary>**My warp route transfer says it was delivered, but I don’t see it**</summary>
   <div>
     We can debug this with two steps:
