+++ src/components/CosmosMessageDelivered.tsx
@@ -1,7 +1,5 @@
-import { useState } from "react";
-import { chainMetadata } from "@hyperlane-xyz/registry";
-
-import ChainDropdown from './ChainDropdown';
+import { ChainMetadata } from "@hyperlane-xyz/sdk";
+import NonEvmMessageDelivered, { strip0x } from "./NonEvmMessageDelivered";
 
 // TODO: these should be in the registry, but for now we'll hardcode them.
 // Once they're in the registry, we can move away from this.
@@ -10,114 +8,37 @@ const mailboxes = {
   'injective': 'inj1palm2wtp6urg0c6j4f2ukv5u5ahdcrqek0sapt',
 };
 
-const cosmosChains = Object.keys(mailboxes);
-
-// Adding @hyperlane-xyz/utils as a dependency breaks things, so we copy the function here.
-export function strip0x(hexstr: string) {
-  return hexstr.startsWith('0x') ? hexstr.slice(2) : hexstr;
+async function messageIsDelivered(metadata: ChainMetadata, mailbox: string, messageId: string): Promise<boolean> {
+  const restUrl = metadata.restUrls?.[0]?.http;
+  if (!restUrl) {
+    throw new Error('No available rest API URL set');
+  }
+  const payload = {
+    mailbox: {
+      message_delivered: {
+        id: strip0x(messageId),
+      }
+    }
+  };
+  const base64Payload = window.btoa(JSON.stringify(payload));
+  const url = `${restUrl}/cosmwasm/wasm/v1/contract/${mailbox}/smart/${base64Payload}`;
+  console.log(`Fetching from ${url}`);
+  const response = await fetch(url, {
+    method: 'GET',
+    headers: {
+      'Content-Type': 'application/json',
+    },
+  });
+  const responseJson = await response.json();
+  return responseJson?.data?.delivered ?? false;
 }
 
 export default function CosmosMessageDelivered({
 }) {
-  const [destinationChain, setDestinationChain] = useState<string>(cosmosChains[0]);
-  const [messageId, setMessageId] = useState('');
-  const [status, setStatus] = useState('');
-
-  const onButtonClick = async () => {
-    const strippedMessageId = strip0x(messageId);
-    if (strippedMessageId.length !== 64) {
-      setStatus(`⛔️ Invalid message ID, must be 32 bytes (64 hex characters)`);
-      return;
-    }
-
-    const metadata = chainMetadata[destinationChain];
-    const mailbox = mailboxes[destinationChain];
-    if (!mailbox) {
-      setStatus(`⛔️ No known Mailbox found for chain ${destinationChain}`);
-      return;
-    }
-    if (!metadata) {
-      setStatus(`⛔️ No known Mailbox found for chain ${destinationChain}`);
-      return;
-    }
-    const restUrl = metadata.restUrls?.[0]?.http;
-    if (!restUrl) {
-      setStatus(`⛔️ No available API set for chain ${destinationChain}`);
-      return;
-    }
-    const payload = {
-      mailbox: {
-        message_delivered: {
-          id: strippedMessageId,
-        }
-      }
-    };
-    const base64Payload = window.btoa(JSON.stringify(payload));
-    const url = `${restUrl}/cosmwasm/wasm/v1/contract/${mailbox}/smart/${base64Payload}`;
-    setStatus(`⏳ Checking if message is delivered...`);
-    let responseJson;
-    console.log(`Fetching from ${url}`);
-    try {
-      const response = await fetch(url, {
-        method: 'GET',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-      });
-      responseJson = await response.json();
-    } catch (e) {
-      setStatus(`⛔️ Error checking message delivery: ${e}`);
-      return;
-    }
-    const delivered = responseJson?.data?.delivered ?? false;
-    if (delivered) {
-      setStatus(`Message successfully delivered: ✅`);
-    } else {
-      setStatus(`Message not yet delivered: ❌`);
-    }
-  };
-
   return (
-    <div style={{
-      padding: "10px",
-      borderRadius: "5px",
-      border: "1px solid #ccc",
-    }}>
-        <ChainDropdown
-          chain={destinationChain}
-          chains={cosmosChains}
-          label="Destination Chain"
-          onChange={setDestinationChain}
-        />
-        <div>
-          Message ID:{"\t"}
-          <input
-            defaultValue={messageId}
-            onChange={(e) => setMessageId(e.target.value)}
-            placeholder="0x..."
-            style={{
-              width: "100%",
-              padding: "5px",
-              borderRadius: "5px",
-              border: "1px solid #ccc",
-              fontSize: "1.05em",
-            }}
-          />
-      </div>
-      <button
-        onClick={onButtonClick}
-        style={{
-          padding: "10px",
-          borderRadius: "5px",
-          border: "1px solid #ccc",
-          marginTop: "10px",
-          fontSize: "1em",
-        }}
-        className="button button--secondary"
-      >Check if delivered</button>
-      <div style={{ margin: "10px" }}>
-        {status}
-      </div>
-    </div>
-  );
+    <NonEvmMessageDelivered
+      mailboxes={mailboxes}
+      messageIsDelivered={messageIsDelivered}
+    />
+  )
 }
